app/requirements.txt
beautifulsoup4==4.12.3
tensorflow==2.16.2
numpy==1.26.4
requests==2.32.2
pillow==10.3.0
Flask==3.0.2
gunicorn==21.2.0
